Some changes in pom's files. Field "Role" added to User and removed from Client. Admin's salary's type chenged to Integer

+++ database/src/main/java/com/loiko/alex/user/AdminDto.java
@@ -17,6 +17,6 @@ import lombok.ToString;
 @ToString(callSuper = true)
 public class AdminDto {
 
-    private Double salary;
+    private Integer salary;
     private AdminRole adminRole;
 }
